HW6. Add rocket, laser and field weapons.

+++ Assets/Scripts/Enemy/EnemyData.cs
@@ -5,7 +5,7 @@ namespace Enemy
     public class EnemyData
     {
         private EnemyView m_View;
-        private int m_Health;
+        private float m_Health;
         
         public EnemyView View => m_View;
         public readonly EnemyAsset Asset;
@@ -22,7 +22,7 @@ namespace Enemy
             m_View.AttachData(this);
         }
 
-        public void GetDamage(int damage)
+        public void GetDamage(float damage)
         {
             m_Health -= damage;
             if (m_Health < 0)
